requirements.txt
mcp>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
